Added event save to DB

+++ app/src/main/java/istvan_torok/agenda/Entities/Event.java
@@ -15,7 +15,7 @@ public class Event {
         this.id = pID;
     }
 
-    public Event(String pDescription, Integer pDate) {
+    public Event(String pDescription, Long pDate) {
         this.description = pDescription;
         this.date = pDate;
     }
